LPS-66931 Move users admin related data handlers to the proper packages
