+++ modules/apps/foundation/users-admin/users-admin-impl/src/main/java/com/liferay/users/admin/internal/exportimport/data/handler/UsersAdminPortletDataHandler.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.users.admin.internal.lar;
+package com.liferay.users.admin.internal.exportimport.data.handler;
 
 import com.liferay.exportimport.kernel.lar.BasePortletDataHandler;
 import com.liferay.exportimport.kernel.lar.DataLevel;
